backend/requirements.txt
fastapi[all]==0.111.0
pydantic-settings==2.3.0
apscheduler==3.10.4
motor==3.4.0
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0
flamapy-smt==1.3.3
neo4j==5.20.0
univers==30.11.0
xmltodict==0.13.0
GitPython==3.1.43
toml==0.10.2
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0